% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab3_filter_utils_apply.R
\name{apply_factor_filter}
\alias{apply_factor_filter}
\title{apply_factor_filter}
\usage{
apply_factor_filter(df, col, vals)
}
\arguments{
\item{df}{A data frame}

\item{col}{A column name, of the factor variable}

\item{vals}{A list of values that were selected}
}
\value{
A filtered data frame
}
\description{
Applying filters to factor variable
}
\examples{
data <- data.frame(fctvar = as.factor(rep(c("a", "b", "c"), 10)))
df_filtered <- apply_factor_filter(data, 'fctvar', c('a', 'b'))

}
