% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_countfactor.R
\name{countfactor}
\alias{countfactor}
\title{countfactor}
\usage{
countfactor(data, var, varlabel)
}
\arguments{
\item{data}{A data frame that contains a factor variable}

\item{var}{The variable name of the factor}

\item{varlabel}{The variable label of the factor}
}
\value{
A data frame that count differentt levels of a factor with four variables: var_label, var_level, freq, and percent
}
\description{
A utils function that counts by the levels of a factor variable
}
\examples{
df <- data.frame(fctvar = as.factor(rep(c("a", "b", "c"), 10)))
countfactor(data = df, var = fctvar, varlabel = 'variable label')
}
