% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visplotly.R
\name{plotly_bar_multifct}
\alias{plotly_bar_multifct}
\title{plotly_bar_multifct}
\usage{
plotly_bar_multifct(data)
}
\arguments{
\item{data}{A data frame generated with [countfactor] function, including 4 variables var_label, var_level, freq, and percent}
}
\value{
A plotly bar chart for multiple factors with same level.
}
\description{
A function to generate a plotly bar chart that combined multiple factors with same levels.
}
\examples{
bar_multifct_df <- data.frame(var_label = c("AA", "AA", "BB", "BB", "CC","CC", "DD", "DD"),
                              var_level = as.factor(rep(c("No", "Yes"), 4)),
                              freq      = c(0, 10, 3, 7, 2, 8, 4, 6),
                              percent   = c("0\%", "100\%", "30\%", "70\%","20\%", "80\%", "40\%", "60\%"))
plotly_bar_multifct(bar_multifct_df)
}
