% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visplotly.R
\name{plotly_histogram_grouped}
\alias{plotly_histogram_grouped}
\title{plotly_histogram_grouped}
\usage{
plotly_histogram_grouped(df, var, groupvar, varlabel, grouplabel)
}
\arguments{
\item{df}{A data frame to be processed}

\item{var}{A numeric variable to generate histogram}

\item{groupvar}{A grouping variable}

\item{label}{Legend label}
}
\value{
A plotly histogram compare a numeric variable between groups.
}
\description{
A standardised function to generate a grouped histogram with three display options.
}
\examples{
group_his_df <- data.frame(numvar = rep(1: 20, 5),
                           fctvar = as.factor(c(rep(c("A", "B", "C"), 30), rep("A", 10))))
plotly_histogram_grouped(df = group_his_df,
                         var = "numvar",
                         groupvar = "fctvar",
                         varlabel = "Numbers",
                         grouplabel = "Groups")

}
