% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visplotly.R
\name{plotly_pie}
\alias{plotly_pie}
\title{plotly_pie}
\usage{
plotly_pie(data, var)
}
\arguments{
\item{data}{A data frame to be processed}

\item{var}{A variable to generate pie chart}
}
\value{
A plotly pie chart.
}
\description{
A standardised function to generate a plotly pie chart
}
\examples{
data <- data.frame(fctvar = as.factor(rep(c("a", "b", "c", "c"), 10)))
plotly_pie(data, fctvar)
}
