% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab3_filter_utils_apply.R
\name{apply_factor_filter}
\alias{apply_factor_filter}
\title{Applying filters to factor variable}
\usage{
apply_factor_filter(df, col, vals)
}
\arguments{
\item{df}{A data frame}

\item{col}{A column name, of the factor variable}

\item{vals}{A list of values that were selected}
}
\value{
A filtered data frame
}
\description{
Applying filters to factor variable
}
\examples{
dfsample <- data.frame(fctvar = as.factor(sample(letters[1:3], 10, replace = TRUE)))
df_filtered <- apply_factor_filter(dfsample, 'fctvar', c('a', 'b'))

}
