% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab3_filter_utils_apply.R
\name{apply_slide_filter}
\alias{apply_slide_filter}
\title{apply_slide_filter}
\usage{
apply_slide_filter(df, col, vals)
}
\arguments{
\item{df}{A data frame}

\item{col}{A column name, of the numeric variable}

\item{vals}{A list of values, min, max, include missing}
}
\value{
A filtered data frame
}
\description{
Applying filters to numeric variable based on the slider inputs
}
\examples{
data <- data.frame(numvar  = rep(c(1,2,3, NA), 10))
df_filtered_a <- apply_slide_filter(data, 'numvar', list(1, 2, 1))
df_filtered_b <- apply_slide_filter(data, 'numvar', list(1, 2, 0))

}
